# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Wedding Invitation API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/galery": {
            "post": {
                "tags": [
                    "Galery"
                ],
                "summary": "Create a new galery",
                "operationId": "b856df5c00cdf6e26611e1930e6eee9f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "undangan_id",
                                    "image"
                                ],
                                "properties": {
                                    "undangan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Galery created successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Undangan not found or unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/undangan/{undangan_id}/galery": {
            "get": {
                "tags": [
                    "Galery"
                ],
                "summary": "Get galery by undangan ID",
                "operationId": "149ff3600b836c465aea4f81a0ec20b9",
                "parameters": [
                    {
                        "name": "undangan_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Galery retrieved successfully"
                    },
                    "404": {
                        "description": "No galery found for this invitation"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/galery/{galery}": {
            "post": {
                "tags": [
                    "Galery"
                ],
                "summary": "Update an existing galery",
                "operationId": "985f4e5f1b677dbcd9ae5aab0d481c8f",
                "parameters": [
                    {
                        "name": "galery",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "undangan_id",
                                    "image"
                                ],
                                "properties": {
                                    "undangan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "type": "file",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Galery updated successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Undangan not found or unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/galery/{id}": {
            "delete": {
                "tags": [
                    "Galery"
                ],
                "summary": "Delete an existing galery",
                "operationId": "a3cf3432d4119e1e77dfeccbe8cdd965",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Galery deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Undangan not found or unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}